build: fix output of pid values
Fix a few more places where pid_t is printed with wrong type
(cherry picked from commit ca1bf4bd7433f05862a6c46d7da4dacaa6a89e2b)

+++ src/util/command.c
@@ -2399,7 +2399,7 @@ void
 virPidAbort(pid_t pid)
 {
     /* Not yet ported to mingw.  Any volunteers?  */
-    VIR_DEBUG("failed to reap child %d, abandoning it", pid);
+    VIR_DEBUG("failed to reap child %lld, abandoning it", (long long)pid);
 }
 
 void
